Reorganise setup into a folder
Looking at the tests we going to find it easier if we split the fetch from the transform. But that means adding another service which is specific only to setup so now seems a good time to move things into a /setup folder.
