+++ app/services/return-requirements/setup/setup.service.js
@@ -5,8 +5,8 @@
  * @module SetupService
  */
 
-const SetupPresenter = require('../../presenters/return-requirements/setup.presenter.js')
-const SessionModel = require('../../models/session.model.js')
+const SetupPresenter = require('../../../presenters/return-requirements/setup.presenter.js')
+const SessionModel = require('../../../models/session.model.js')
 
 /**
  * Orchestrates fetching and presenting the data for `/return-requirements/{sessionId}/setup` page
