+++ test/controllers/return-requirements.controller.test.js
@@ -25,7 +25,7 @@ const RemoveService = require('../../app/services/return-requirements/remove.ser
 const ReturnCycleService = require('../../app/services/return-requirements/returns-cycle.service.js')
 const SelectPurposeService = require('../../app/services/return-requirements/purpose.service.js')
 const SelectReasonService = require('../../app/services/return-requirements/reason.service.js')
-const SetupService = require('../../app/services/return-requirements/setup.service.js')
+const SetupService = require('../../app/services/return-requirements/setup/setup.service.js')
 const SiteDescriptionService = require('../../app/services/return-requirements/site-description.service.js')
 const StartDateService = require('../../app/services/return-requirements/start-date.service.js')
 const SubmitAbstractionPeriod = require('../../app/services/return-requirements/submit-abstraction-period.service.js')
@@ -38,7 +38,7 @@ const SubmitPointsService = require('../../app/services/return-requirements/subm
 const SubmitPurposeService = require('../../app/services/return-requirements/submit-purpose.service.js')
 const SubmitReasonService = require('../../app/services/return-requirements/submit-reason.service.js')
 const SubmitReturnsCycleService = require('../../app/services/return-requirements/submit-returns-cycle.service.js')
-const SubmitSetupService = require('../../app/services/return-requirements/submit-setup.service.js')
+const SubmitSetupService = require('../../app/services/return-requirements/setup/submit-setup.service.js')
 const SubmitSiteDescriptionService = require('../../app/services/return-requirements/submit-site-description.service.js')
 const SubmitStartDateService = require('../../app/services/return-requirements/submit-start-date.service.js')
